rename lstm_model->ml_model
