+++ experiment/tools/strength_test.py
